renamed nutrient namespace to nutrients
