+++ Models/Core/Model.cs
@@ -110,12 +110,12 @@ namespace Models.Core
         [XmlElement(typeof(Soils.OutputLayers))]
         [XmlElement(typeof(Soils.Arbitrator.SoilArbitrator))]
         [XmlElement(typeof(Soils.Sample))]
-        [XmlElement(typeof(Soils.Nutrient.Nutrient))]
-        [XmlElement(typeof(Soils.Nutrient.NutrientPool))]
-        [XmlElement(typeof(Soils.Nutrient.CarbonFlow))]
-        [XmlElement(typeof(Soils.Nutrient.NFlow))]
-        [XmlElement(typeof(Soils.Nutrient.Solute))]
-        [XmlElement(typeof(Soils.Nutrient.Chloride))]
+        [XmlElement(typeof(Soils.Nutrients.Nutrient))]
+        [XmlElement(typeof(Soils.Nutrients.NutrientPool))]
+        [XmlElement(typeof(Soils.Nutrients.CarbonFlow))]
+        [XmlElement(typeof(Soils.Nutrients.NFlow))]
+        [XmlElement(typeof(Soils.Nutrients.Solute))]
+        [XmlElement(typeof(Soils.Nutrients.Chloride))]
         [XmlElement(typeof(WaterModel.CNReductionForCover))]
         [XmlElement(typeof(WaterModel.CNReductionForTillage))]
         [XmlElement(typeof(WaterModel.EvaporationModel))]
