+++ Models/Models.csproj
@@ -216,7 +216,7 @@
     <Compile Include="Plant\Phenology\Responses\Vernalisation.cs" />
     <Compile Include="Plant\Phenology\Scales\BBCH.cs" />
     <Compile Include="Plant\Structure\BudNumberFunction.cs" />
-    <Compile Include="Soils\Nutrient\Chloride.cs" />
+    <Compile Include="Soils\Nutrients\Chloride.cs" />
     <Compile Include="Plant\Structure\ApexBase.cs" />
     <Compile Include="Storage\DataStore.cs" />
     <Compile Include="Graph\GraphPage.cs" />
@@ -331,17 +331,17 @@
     <Compile Include="Report\ReportColumnWithValues.cs" />
     <Compile Include="Report\ReportColumnConstantValue.cs" />
     <Compile Include="Report\ReportColumn.cs" />
-    <Compile Include="Soils\Nutrient\NFlow.cs" />
-    <Compile Include="Soils\Nutrient\CFlow.cs" />
-    <Compile Include="Soils\Nutrient\NutrientPool.cs" />
-    <Compile Include="Soils\Nutrient\Nutrient.cs" />
+    <Compile Include="Soils\Nutrients\NFlow.cs" />
+    <Compile Include="Soils\Nutrients\CFlow.cs" />
+    <Compile Include="Soils\Nutrients\NutrientPool.cs" />
+    <Compile Include="Soils\Nutrients\Nutrient.cs" />
     <Compile Include="Soils\Arbitrator\CropUptakes.cs" />
     <Compile Include="Soils\Arbitrator\Estimate.cs" />
     <Compile Include="Soils\Arbitrator\SoilState.cs" />
     <Compile Include="Soils\HydraulicProperties.cs" />
     <Compile Include="Soils\ISoilCrop.cs" />
     <Compile Include="Soils\Arbitrator\ZoneWaterAndN.cs" />
-    <Compile Include="Soils\Nutrient\Solute.cs" />
+    <Compile Include="Soils\Nutrients\Solute.cs" />
     <Compile Include="Soils\OutputLayers.cs" />
     <Compile Include="Soils\MultiPoreWater\Evapotranspiration.cs" />
     <Compile Include="Soils\MultiPoreWater\MRSpline.cs" />
