+++ Models/Soils/Nutrients/Chloride.cs
@@ -1,4 +1,4 @@
-﻿namespace Models.Soils.Nutrient
+﻿namespace Models.Soils.Nutrients
 {
     using Core;
     using Interfaces;
