+++ Models/Soils/Nutrients/NFlow.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace Models.Soils.Nutrient
+namespace Models.Soils.Nutrients
 {
     using Core;
     using Models.Functions;
